refactor(QiScans): adopt mixins and merge network files
