+++ src/QiScans/pbconfig.ts
@@ -4,7 +4,7 @@ import { ContentRating, SourceIntents } from "@paperback/types";
 export default {
   name: "QiScans",
   description: "Extension that pulls content from qiscans.org.",
-  version: "1.0.0-alpha.12",
+  version: "1.0.0-alpha.14",
   icon: "icon.png",
   language: "en",
   contentRating: ContentRating.EVERYONE,
