+++ src/QiScans/services/network.ts
@@ -1,5 +1,5 @@
 import type { Request, Response } from "@paperback/types";
-import { PaperbackInterceptor } from "@paperback/types";
+import { CloudflareError, PaperbackInterceptor } from "@paperback/types";
 import { QISCANS_DOMAIN } from "../main";
 
 export class QiScansInterceptor extends PaperbackInterceptor {
@@ -16,5 +16,19 @@ export class QiScansInterceptor extends PaperbackInterceptor {
     data: ArrayBuffer,
   ): Promise<ArrayBuffer> {
     return data;
+  }
+}
+
+export function checkCloudflareStatus(request: Request, status: number): void {
+  if (status === 503) {
+    throw new CloudflareError({
+      url: request.url,
+      method: request.method ?? "GET",
+    });
+  }
+  if (status === 403) {
+    throw new Error(
+      "Server returned 403 Forbidden. This title may have been removed due to a DMCA request or is otherwise unavailable.",
+    );
   }
 }
